Putting the language selector in the app.tsx
