+++ webapp/src/RegisterForm.tsx
@@ -8,35 +8,6 @@ type Props = {
   onGoToSignUp: () => void;
 };
 
-export function LanguageSwitcher() {
-  const { i18n } = useTranslation();
-  const LANGUAGES = [
-    { code: "en", label: "ENG" },
-    { code: "es", label: "ESP" },
-    { code: "fi", label: "FIN" },
-    { code: "tr", label: "TUR" },
-    { code: "as", label: "AST" },
-  ];
-
-  return (
-    <div className="language-container">
-      {LANGUAGES.map((lang) => {
-        const isActive = i18n.language.startsWith(lang.code);
-        return (
-          <button
-            key={lang.code}
-            type="button"
-            onClick={() => i18n.changeLanguage(lang.code)}
-            className={`lang-button ${isActive ? "active" : ""}`}
-          >
-            <span className="lang-label">{lang.label}</span>
-          </button>
-        );
-      })}
-    </div>
-  );
-}
-
 const LoginForm: React.FC<Props> = ({ onLoggedIn, onGoToSignUp }) => {
   const { t } = useTranslation();
   const [username, setUsername] = useState("");
@@ -98,8 +69,6 @@ const LoginForm: React.FC<Props> = ({ onLoggedIn, onGoToSignUp }) => {
 
   return (
     <div className="login-form-container">
-      <LanguageSwitcher />
-
       <form onSubmit={handleSubmit} className="register-form">
         <h2>{t("login.title")}</h2>
         <p>{t("login.subtitle")}</p>
